+++ nimble-rs/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+* (Breaking) Update to bt-hci 0.10; `CoreCmds` no longer requires `Read Local Supported Commands` as nrf-sdc 0.4.0 does not support it anyway
 * Update the README.md to explain the main caveat of the approach and its consequences
 
 ## [0.1.0] - 2026-08-20